+++ src/QNMDNS.cpp
@@ -46,14 +46,15 @@ static void srv_txt(struct mdns_service *service, void *txt_userdata) {
 }
 
 static bool initialized = false;
+static bool netifAdded = false;
 
 MDNSClass::~MDNSClass() {
   end();
 }
 
 bool MDNSClass::begin(const char *hostname) {
-  netif_ = netif_default;
-  if (netif_ == nullptr) {
+  if (netif_default == nullptr) {
+    // Return false for no netif
     return false;
   }
 
@@ -62,27 +63,32 @@ bool MDNSClass::begin(const char *hostname) {
     initialized = true;
   }
 
-  if (mdns_resp_add_netif(netif_, hostname, kTTL) != ERR_OK) {
-    netif_ = nullptr;
+  if (netifAdded) {
     return false;
   }
+  if (mdns_resp_add_netif(netif_default, hostname, kTTL) != ERR_OK) {
+    return false;
+  }
+  netifAdded = true;
+  netif_ = netif_default;
   hostname_ = hostname;
   return true;
 }
 
 bool MDNSClass::end() {
-  if (netif_ == nullptr) {
+  if (!netifAdded) {
     // Return true for no netif
     return true;
   }
   bool retval = (mdns_resp_remove_netif(netif_) == ERR_OK);
+  netifAdded = false;
   netif_ = nullptr;
   hostname_ = "";
   return retval;
 }
 
 void MDNSClass::restart() {
-  if (netif_ == nullptr) {
+  if (!netifAdded) {
     return;
   }
   mdns_resp_restart(netif_);
@@ -117,7 +123,8 @@ bool MDNSClass::addService(const char *type, const char *protocol,
 bool MDNSClass::addService(const char *name, const char *type,
                            const char *protocol, uint16_t port,
                            std::vector<String> (*getTXTFunc)(void)) {
-  if (netif_ == nullptr) {
+  if (!netifAdded) {
+    // Return false for no netif
     return false;
   }
 
@@ -155,9 +162,9 @@ bool MDNSClass::removeService(const char *type, const char *protocol,
 
 bool MDNSClass::removeService(const char *name, const char *type,
                               const char *protocol, uint16_t port) {
-  if (netif_ == nullptr) {
-    // Return true for no netif
-    return true;
+  if (!netifAdded) {
+    // Return false for no netif
+    return false;
   }
 
   // Find a matching service
@@ -176,7 +183,7 @@ uint32_t MDNSClass::ttl() const {
 }
 
 void MDNSClass::announce() const {
-  if (netif_ == nullptr) {
+  if (!netifAdded) {
     return;
   }
   mdns_resp_announce(netif_);